+++ roxctl/common/tls.go
@@ -11,41 +11,42 @@ import (
 	"github.com/stackrox/rox/pkg/clientconn"
 	"github.com/stackrox/rox/pkg/mtls"
 	"github.com/stackrox/rox/pkg/netutil"
-	"github.com/stackrox/rox/pkg/sync"
 	"github.com/stackrox/rox/pkg/tlscheck"
 	"github.com/stackrox/rox/roxctl/common/flags"
 	"github.com/stackrox/rox/roxctl/common/logger"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
-const warningMsg = `The remote endpoint failed TLS validation. This will be a fatal error in future releases.
-Please do one of the following at your earliest convenience:
+const errorMsg = `The remote endpoint failed TLS validation.
+  Please do one of the following:
   1. Obtain a valid certificate for your Central instance/Load Balancer.
-  2. Use the --ca option to specify a custom CA certificate (PEM format). This Certificate can be obtained by
-     running "roxctl central cert".
-  3. Update all your roxctl usages to pass the --insecure-skip-tls-verify option, in order to
-     suppress this warning and retain the old behavior of not validating TLS certificates in
-     the future (NOT RECOMMENDED).
-
+  2. Use the --ca option to specify a custom CA certificate (PEM format).
+     This certificate can be obtained by running "roxctl central cert".
+  3. Use the --insecure-skip-tls-verify option to suppress this error
+     and not validate TLS certificates (NOT RECOMMENDED).
 `
 
-type insecureVerifierWithWarning struct {
-	printWarningOnce sync.Once
-	logger           logger.Logger
+type grpcPermissionDenied struct{ error }
+
+func (p *grpcPermissionDenied) GRPCStatus() *status.Status {
+	return status.New(codes.PermissionDenied, p.Error())
 }
 
-func (v *insecureVerifierWithWarning) VerifyPeerCertificate(leaf *x509.Certificate, chainRest []*x509.Certificate, conf *tls.Config) error {
+type insecureVerifierWithError struct {
+	logger logger.Logger
+}
+
+func (v *insecureVerifierWithError) VerifyPeerCertificate(leaf *x509.Certificate, chainRest []*x509.Certificate, conf *tls.Config) error {
 	verifyOpts := x509.VerifyOptions{
 		DNSName:       conf.ServerName,
 		Intermediates: tlscheck.NewCertPool(chainRest...),
 		Roots:         conf.RootCAs,
 	}
 
-	_, err := leaf.Verify(verifyOpts)
-	if err != nil {
-		v.printWarningOnce.Do(func() {
-			v.logger.WarnfLn(warningMsg)
-			v.logger.WarnfLn("Certificate validation error: %v", err)
-		})
+	if _, err := leaf.Verify(verifyOpts); err != nil {
+		v.logger.ErrfLn(errorMsg)
+		return &grpcPermissionDenied{err}
 	}
 	return nil
 }
@@ -106,7 +107,7 @@ func tlsConfigOptsForCentral(logger logger.Logger) (*clientconn.TLSConfigOptions
 				return nil, errors.Wrap(err, "failed to parse CA certificates from file")
 			}
 		} else {
-			customVerifier = &insecureVerifierWithWarning{
+			customVerifier = &insecureVerifierWithError{
 				logger: logger,
 			}
 			// Read the CA from the central secret.