+++ tests/e2e/lib.sh
@@ -34,6 +34,8 @@ deploy_stackrox() {
 
     export_central_basic_auth_creds
     wait_for_api "${central_namespace}"
+    export_central_cert
+
     setup_client_TLS_certs "${tls_client_certs}"
     record_build_info "${central_namespace}"
 
@@ -363,7 +365,7 @@ deploy_sensor_via_operator() {
     local central_endpoint="central.${central_namespace}.svc:443"
 
     info "Deploying sensor via operator into namespace ${sensor_namespace} (central is expected in namespace ${central_namespace})"
-
+    export_central_cert
     if ! kubectl get ns "${sensor_namespace}" >/dev/null 2>&1; then
         kubectl create ns "${sensor_namespace}"
     fi
@@ -452,6 +454,23 @@ export_central_basic_auth_creds() {
     ROX_USERNAME="admin"
     ci_export "ROX_USERNAME" "$ROX_USERNAME"
     ci_export "ROX_PASSWORD" "$ROX_PASSWORD"
+}
+
+export_central_cert() {
+    if [[ -f "${ROX_CA_CERT_FILE:-}" ]]; then
+        info "(e2e) Using central certificate from ${ROX_CA_CERT_FILE} ($(md5sum "${ROX_CA_CERT_FILE}"))"
+        openssl x509 -in "${ROX_CA_CERT_FILE}" -subject -issuer -noout
+        return
+    fi
+    require_environment "API_ENDPOINT"
+    require_environment "ROX_PASSWORD"
+
+    ci_export ROX_CA_CERT_FILE "$(mktemp -d)/central_cert.pem"
+    info "Storing central certificate in ${ROX_CA_CERT_FILE}"
+
+    roxctl -e "$API_ENDPOINT" -p "$ROX_PASSWORD" \
+        central cert --insecure-skip-tls-verify 1>"$ROX_CA_CERT_FILE"
+    info "md5sum $(md5sum "${ROX_CA_CERT_FILE}")"
 }
 
 deploy_optional_e2e_components() {
@@ -1164,6 +1183,7 @@ db_backup_and_restore_test() {
 
     # Ensure central is ready for requests after any previous tests
     wait_for_api "${central_namespace}"
+    export_central_cert
 
     info "Backing up to ${output_dir}"
     mkdir -p "$output_dir"